requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Digital Ocean Spaces (S3 compatible)
boto3>=1.34.0

# ML & Data
joblib>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Streamlit
streamlit>=1.29.0

# OpenAI
openai>=1.6.0

# Langfuse
langfuse>=2.6.0

# Utilities
python-dotenv>=1.0.0
